My lab2 answer

+++ lab2/DebugExercise/DebugExercise1.java
@@ -5,10 +5,10 @@ package DebugExercise;
  * feature work.
  */
 public class DebugExercise1 {
-    public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
-        int result = Math.round(quotient);
-        return result;
+    public static int divideThenRound(double top, double bottom) {
+        double quotient = top / bottom;
+        double result = Math.round(quotient);
+        return (int)result;
     }
 
     public static void main(String[] args) {
